# Reimplement decode so it takes AsRef<[u8]>