This is a breaking change.

`bubblebabble::decode` is able to take `AsRef<[u8]>` by using `bstr` to
assert that the slice only contains ASCII bytes.

This PR changes the errors returned by `decode` in some circumstances
when the given byte slice is less than 5 ASCII characters long.

- [BREAKING] `DecodeError::InvalidSymbol` takes the index in the byte
  slice where the first invalid symbol occured instead of a `char`.
- [BREAKING] `DecodeError::NonAscii` takse the index in the byte slice
  where the first non-ASCII byte occurred instead of a `char`.
- Use slice patterns to destructure header and trailer out of the
  encoded byte slice.
- `checksum` is now a `u8` that is temporarily promoted to `u16` when
  updating the checksum at the end of each loop iteration.
- Add documentation around overflow assumptions.
- Remove use of `u8::try_from` now that `checksum` is a `u8`.