@lopopolo 2020-03-20
# Performance

This PR improves `decode` performance by 8%.

## Before, master @ 4205291

https://github.com/artichoke/bubblebabble/runs/521113709?check_suite_focus=true

```
test decode::benchmark_decode_empty             ... bench:           3 ns/iter (+/- 0)
test decode::benchmark_decode_vector_1234567890 ... bench:         169 ns/iter (+/- 39)
test decode::benchmark_decode_vector_pineapple  ... bench:         160 ns/iter (+/- 43)
test encode::benchmark_encode_empty             ... bench:          19 ns/iter (+/- 3)
test encode::benchmark_encode_vector_1234567890 ... bench:         101 ns/iter (+/- 32)
test encode::benchmark_encode_vector_pineapple  ... bench:          80 ns/iter (+/- 13)
```

## After, GH-20 @  96fb776

https://github.com/artichoke/bubblebabble/runs/522224564?check_suite_focus=true

```
test decode::benchmark_decode_empty             ... bench:           4 ns/iter (+/- 0)
test decode::benchmark_decode_vector_1234567890 ... bench:         149 ns/iter (+/- 13)
test decode::benchmark_decode_vector_pineapple  ... bench:         130 ns/iter (+/- 12)
test encode::benchmark_encode_empty             ... bench:          23 ns/iter (+/- 4)
test encode::benchmark_encode_vector_1234567890 ... bench:         118 ns/iter (+/- 15)
test encode::benchmark_encode_vector_pineapple  ... bench:         104 ns/iter (+/- 12)
```